Revamp how settings are defined, found & stored (#360)
* Revamp how settings are defined, found & stored

Added a `Setting` object which defines information related to each setting (description, name, alternate names, default value, etc).  Settings can even define a dynamic source for it’s default value. This will ensure that we always know everything about each setting we define and where/how the default values are determined.  Removes a lot of parameter/setting copying code to make sure things are named right and in the right properties instance.

Allowing alternate names for a setting allows changing them in the future. This isn’t just a “nice to have” currently we use incorrect names for datasource properties; this allows a backwards compatible fix.

Also, added `Settings` which is much like `java.util.Properties` in that it stores everything in text but it allows access via `Setting` definitions.  `Settings` logs warnings for unrecognized names and logs errors for invalid values that are stored.

* Rewrite abstract data source using new settings

* `AbstractDataSource` now stores everything in a `Settings` bag that recognizes JDBC, JDBC-DS, SYS & PROTO setting groups.
  * Unit test was added to ensure all properties from these groups are available as JavaBean properties on the data source.
* Added support for `databaseUrl` to `AbstractDataSource`
* All settings now have a “dotted” primary name (enforced by `Setting` class) this ensures no capitalization is required for properties and the name can be easily made into a JavaBean property name.

* Add literal escaping methods to SQLTextUtils

* Update application_name via parameter status

* Ensure settings knows required JDBC/System groups

* Implement connection client-info via settings

Implements client-info metadata & actions via settings.  Supports `ApplicationName` & `ClientUser`.

* Implement Driver.getPropertyInfo via Settings

Now lists all available settings including descriptions.

+++ src/test/java/com/impossibl/postgres/jdbc/StatementTest.java
@@ -597,10 +597,10 @@ public class StatementTest {
 
   @Test
   public void testDefaultFetchSize() throws SQLException {
-    int oldValue = ((PGConnection) con).getDefaultFetchSize();
+    Integer oldValue = ((PGConnection) con).getDefaultFetchSize();
 
     ((PGConnection)con).setDefaultFetchSize(10);
-    assertEquals(10, ((PGConnection)con).getDefaultFetchSize());
+    assertEquals((Integer) 10, ((PGConnection)con).getDefaultFetchSize());
 
     Statement stmt = con.createStatement();
     assertEquals(10, stmt.getFetchSize());
